-Possibility to dispatch customState -added remove and clear callback from Process -August checkpoint

+++ src/main/java/com/istat/freedev/processor/ProcessManager.java
@@ -364,7 +364,7 @@ public final class ProcessManager {
 
     }
 
-    public final void postDelayed(Runnable runnable, int delayed) {
+    public final void postDelayed(Runnable runnable, long delayed) {
         mHandler.postDelayed(runnable, delayed);
     }
 }
